pyproject.toml
[tool.poetry]
name = "icns_gen"
version = "0.8.0"
description = "Create a macos icns file."
license = "MIT"
authors = ["bubio <bubio66@gmail.com>"]
repository = "https://github.com/bubio/icns_gen.git"

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^8.1.1" 

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"